refactor(backend): move test-id codecs out of the graphql layer
The test/change ID codecs and the fingerprint codec lived under
graphql/services, so the REST API had to import from the graphql layer. Move
them to @/util so both the API and GraphQL depend on a transport-neutral module.
