+++ apps/backend/src/util/fingerprint.ts
